requirements.txt
# Shareholder Catalyst - Python Dependencies

# Core APIs
requests>=2.31.0
openai>=1.0.0

# Market Data
yfinance>=0.2.0

# Environment Variables
python-dotenv>=1.0.0

# Web Interface
streamlit>=1.28.0

# PDF Generation
reportlab>=4.0.0

# Note: Using LandingAI Direct API (no SDK required)
